ServerSage/requirements.txt
google-generativeai>=0.3.0
discord.py>=2.0.0
Red-DiscordBot>=3.5.0
aiohttp>=3.8.0
